+++ Dockerfile
@@ -12,6 +12,9 @@ ENV VC_TIMEOUT=
 ENV MAX_SONG_PRELOAD=
 ENV HEROKU=
 ENV ENABLE_SLASH_COMMANDS=
+ENV VC_TIMOUT_DEFAULT=
+ENV MAX_HISTORY_LENGTH=
+ENV MAX_TRACKNAME_HISTORY_LENGTH=
 
 RUN pip --no-cache-dir install -r requirements.txt \
     && apt-get update \
